+++ scripts/release.py
@@ -5,6 +5,8 @@
 # ]
 # ///
 import glob
+import hashlib
+import os
 import subprocess
 
 import typer
@@ -49,16 +51,41 @@ def release(dry_run: bool = False):
         if line.startswith("## "):
             break
     changes = "".join(changes[: i - 1])
+    changes += """
+---
+
+You can find the Python packages on [PyPI](https://pypi.org/project/redpepper/).
+
+The `checksums.txt` file contains SHA256 checksums for the release assets.
+"""
     typer.secho(changes, fg=typer.colors.YELLOW)
 
     typer.secho(
-        "Will create Github release with these artifacts:",
+        "Will create Github release with these assets:",
         fg=typer.colors.BLUE,
         bold=True,
     )
     files = glob.glob("dist/*.whl") + glob.glob("dist/*.tar.gz")
+    checksums = {}
     for file in files:
         typer.echo(f"  - {typer.style(file, fg=typer.colors.MAGENTA)}")
+        with open(file, "rb") as f:
+            hash = hashlib.sha256()
+            while chunk := f.read(4096):
+                hash.update(chunk)
+            checksums[os.path.basename(file)] = hash.hexdigest()
+
+    if not dry_run:
+        with open("dist/checksums.txt", "w") as f:
+            for file, checksum in checksums.items():
+                f.write(f"{checksum}  {file}\n")
+    else:
+        typer.secho("Checksums:", fg=typer.colors.BLUE, bold=True)
+        for file, checksum in checksums.items():
+            typer.echo(f"{checksum}  {file}")
+
+    files.append("dist/checksums.txt")
+    typer.echo(f"  - {typer.style('dist/checksums.txt', fg=typer.colors.MAGENTA)}")
 
     if not typer.confirm(
         typer.style("OK to proceed?", fg=typer.colors.BRIGHT_CYAN, bold=True),
@@ -88,6 +115,8 @@ def release(dry_run: bool = False):
             != 0
         ):
             raise typer.Abort()
+    else:
+        typer.echo("Would create Github release here")
 
     typer.secho("All done!", fg=typer.colors.GREEN)